refactor: move Service related to Word class to workanalysis package
Moved 'MorphemeApiClient', 'MorphemeExtractionService',
'WordExtractionService'
from 'com.example.spinlog.statistics.service'
to 'com.example.spinlog.statistics.service.wordanalysis'
